Moved namespaces to a separate directory
