--- library/namespaces/mod.ts
@@ -1,19 +1,0 @@
-export { dbo } from "./dbo.ts";
-export { dc } from "./dc.ts";
-export { dcterms } from "./dcterms.ts";
-export { foaf } from "./foaf.ts";
-export { gr } from "./gr.ts";
-export { ldkit } from "./ldkit.ts";
-export { owl } from "./owl.ts";
-export { rdf } from "./rdf.ts";
-export { rdfs } from "./rdfs.ts";
-export { schema } from "./schema.ts";
-export { sioc } from "./sioc.ts";
-export { skos } from "./skos.ts";
-export { xsd } from "./xsd.ts";
-
-export {
-  createNamespace,
-  type Namespace,
-  type NamespaceInterface,
-} from "./namespace.ts";
