+++ namespaces.ts
@@ -1,1 +1,15 @@
-export * from "./library/namespaces/mod.ts";
+export { dbo } from "./namespaces/dbo.ts";
+export { dc } from "./namespaces/dc.ts";
+export { dcterms } from "./namespaces/dcterms.ts";
+export { foaf } from "./namespaces/foaf.ts";
+export { gr } from "./namespaces/gr.ts";
+export { ldkit } from "./namespaces/ldkit.ts";
+export { owl } from "./namespaces/owl.ts";
+export { rdf } from "./namespaces/rdf.ts";
+export { rdfs } from "./namespaces/rdfs.ts";
+export { schema } from "./namespaces/schema.ts";
+export { sioc } from "./namespaces/sioc.ts";
+export { skos } from "./namespaces/skos.ts";
+export { xsd } from "./namespaces/xsd.ts";
+
+export * from "./library/namespace.ts";
