+++ namespaces/rdf.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "./namespace.ts";
+import { createNamespace } from "../library/namespace.ts";
 
 /**
  * The RDF Concepts Vocabulary (RDF)
