+++ namespaces/rdfs.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "./namespace.ts";
+import { createNamespace } from "../library/namespace.ts";
 
 /**
  * The RDF Schema vocabulary (RDFS)
